Drop port_prototpyes.h back to functions.h, consolidate
